Add logic to map GitHub issue with calendar entry
When a calendar entry gets created out of a session issue, the calendar entry
URL now includes a link to the session issue, used both to advertize the issue
and to do a security check when calendar update comment runs.

Conversely, the session issue gets updated afterwards to include a link back to
the calendar entry in its "Materials" section.