Relax validation for session goal and drop-down options
When people edit options afterwards, e.g., to choose an option in the list,
they may introduce changes to the case of characters. For instance,
"20-45 People" instead of "20-45 people".

Similarly, while the session goal is a simple text input, people may edit the
goal afterwards and introduce markdown. That's fine, let's accept that.